The Sentiment on Andrew Mountbatten-Windsor
This code focuses on the sentiment towards Andrew Mountbatten-Windsor using Python to extract sentiment from YouTube comments via an API. I aimed to focus on LBC because it's a radio call-in channel, which makes it more opinion-oriented, especially with the hosts often posing questions to the audience. When I lived in Birmingham and London, I would often see adverts for LBC, so I knew it was a good starting point. The code concentrates on the sentiment regarding the former Prince Andrew and how his latest scandal might influence public opinion on the monarchy and the Met police. Enjoy!